+++ src/main/java/com/ljs/ootp/ai/stats/PitchingStats.java
@@ -32,13 +32,13 @@ public class PitchingStats implements Stats<PitchingStats> {
     public void setAtBats(int atBats) { this.atBats = atBats; }
 
     public Integer getHits() { return hits; }
-    public void setHits(int hits) { this.hits = hits; }
+    public void setHits(long hits) { this.hits = (int) hits; }
 
-    public void setDoubles(int doubles) { this.doubles = doubles; }
-    public void setTriples(int triples) { this.triples = triples; }
-    public void setStrikeouts(int strikeouts) { this.strikeouts = strikeouts; }
-    public void setWalks(int walks) { this.walks = walks; }
-    public void setHomeRuns(int homeRuns) { this.homeRuns = homeRuns; }
+    public void setDoubles(long doubles) { this.doubles = (int) doubles; }
+    public void setTriples(long triples) { this.triples = (int) triples; }
+    public void setStrikeouts(long strikeouts) { this.strikeouts = (int) strikeouts; }
+    public void setWalks(long walks) { this.walks = (int) walks; }
+    public void setHomeRuns(long homeRuns) { this.homeRuns = (int) homeRuns; }
 
     public void setLeaguePitching(PitchingStats leaguePitching) {
         this.leaguePitching = leaguePitching;
